% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsAssess.R
\name{bheqWrapper}
\alias{bheqWrapper}
\title{Calculate the equilibrium Beverton-Holt estimator of instantaneous total mortality from length data. Wrapper function from fishmethods R package.}
\usage{
bheqWrapper(LifeHistoryObj, LengthCompObj, byGroup = FALSE, mode = NULL)
}
\arguments{
\item{LifeHistoryObj}{A LifeHistory object from fishSimGTG.}

\item{LengthCompObj}{A LengthComp object}

\item{byGroup}{A logical indicating whether quantity is to be calculated separately for each of multiple length comp groups (TRUE) or to length comp is to be pooled across groups prior to calculating quantity (default = FALSE). When TRUE, pooling is ignored if only a single group exists.}

\item{mode}{Mode of length-frequency distribution (optional). Used as the length at full recruitment to the fishery. User can apply LcFunc or LcFuncKernel to the length data previously. If no value is supplied, an estimate will be automatically obtained using modeKernelFunc.}
}
\description{
Nelson GA (2023). fishmethods: Fishery Science Methods and Models. R package version 1.12-1, \url{https://CRAN.R-project.org/package=fishmethods}.
}
\examples{
library(fishSimGTG)
library(fishmethods)
bheqWrapper(fishSimGTG::LifeHistoryExample, fishLengthAssess::LengthCompExampleFreq)
}
