% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishLengthAssess.R
\docType{package}
\name{fishLengthAssess}
\alias{fishLengthAssess}
\alias{fishLengthAssess-package}
\title{Length-based evaluation of fish stock and fishery}
\description{
Evaluates stock status and fishery compliance, and conducts other management-related tasks based on length-frequency data.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
