% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_observed_freq}
\alias{get_observed_freq}
\title{Convert LengthComp data to frequency for plotting
Converts either Frequency or Length data from a LengthComp object into
a frequency distribution using the model length bins.
Allows observed vs predicted plots to be created regardless of the original(lengths or frecuency)
data type}
\usage{
get_observed_freq(lengthCompObj, SizeBins)
}
\arguments{
\item{lengthCompObj}{A LengthComp S4 object with dataType "Frequency" or "Length".}

\item{SizeBins}{A list with Linc and ToSize elements (from make_size_bins()).}
}
\value{
A data frame with columns Length and Observed.
}
\description{
Convert LengthComp data to frequency for plotting
Converts either Frequency or Length data from a LengthComp object into
a frequency distribution using the model length bins.
Allows observed vs predicted plots to be created regardless of the original(lengths or frecuency)
data type
}
\examples{
library(fishSimGTG)
SizeBins <- make_size_bins(fishSimGTG::LifeHistoryExample)
obs_df <- get_observed_freq(fishLengthAssess::LengthCompExampleFreq, SizeBins)
}
