% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsAssess.R
\name{lbsprWrapper}
\alias{lbsprWrapper}
\title{Estimation of SPR using LBSPR method (Hordyk et al. 2016)}
\usage{
lbsprWrapper(
  LifeHistoryObj,
  LengthCompObj,
  Lc = 0,
  binWidth = 1,
  cvLinf = 0.1,
  byGroup = FALSE
)
}
\arguments{
\item{LifeHistoryObj}{A LifeHistory object from fishSimGTG.}

\item{LengthCompObj}{A LengthComp object}

\item{Lc}{A minimum size limit in the same units of measure (L_units) and type of measurement (L_type) as the corresponding LengthComp object. Only used for analysis of fishery-independent length composition data.}

\item{binWidth}{Used to create length-frequencies when LengthComp dataType is raw lengths.}

\item{cvLinf}{Variability in Linf, used in LBSPR fitting. Default value same as that used in LBSPR library}

\item{byGroup}{A logical indicating whether quantity is to be calculated separately for each of multiple length comp groups (TRUE) or to length comp is to be pooled across groups prior to calculating quantity (default = FALSE). When TRUE, pooling is ignored if only a single group exists.}
}
\description{
Hordyk, A.R., Ono, K., Prince, J.D., Walters, C.J., 2016. A simple length-structured model based on life history ratios and incorporating size-dependent selectivity: application to spawning potential ratios for data-poor stocks. Can. J. Fish. Aquat. Sci. 73, 1787–1799. https://doi.org/10.1139/cjfas-2015-0422
}
\examples{
library(fishSimGTG)
library(LBSPR)
lbsprWrapper(fishSimGTG::LifeHistoryExample, fishLengthAssess::LengthCompExample)
}
