% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{LcFunc}
\alias{LcFunc}
\title{Estimate length at full selectivity using the mode of the length-frequency distribution
#'}
\usage{
LcFunc(LengthCompObj, byGroup = FALSE)
}
\arguments{
\item{LengthCompObj}{A LengthComp object.}

\item{byGroup}{A logical indicating whether quantity is to be calculated separately for each of multiple length comp groups (TRUE) or to length comp is to be pooled across groups prior to calculating quantity (default = FALSE). When TRUE, pooling is ignored if only a single group exists.}
}
\description{
Estimate length at full selectivity using the mode of the length-frequency distribution
#'
}
\examples{
library(fishSimGTG)
LcFunc(fishLengthAssess::LengthCompExampleFreq, byGroup = FALSE)
}
