% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesS4.R
\docType{class}
\name{LengthComp-class}
\alias{LengthComp-class}
\title{Length composition object}
\arguments{
\item{title}{A title for the object, useful for displaying the contents of the object}

\item{description}{A longer description of the object}

\item{L_units}{Units of measure for object, e.g., cm}

\item{L_type}{Options are TL, FL, or SL (Total length, fork length, and standard length, respectively)}

\item{dataType}{Options are Length or Frequency (See details for additional information)}

\item{observationUnit}{Provide a description for interpreting column header, e.g., year}

\item{header}{Logical. Whether dataframe contains a descriptive  header. FALSE will cause the header to be hidden}

\item{dt}{The length composition data frame.}
}
\description{
An S4 object that holds length composition data in a specified format.
}
\details{
The length composition object is contains a data frame of length observations.
These observations are organized according to 'dataType'. When 'dataType' is set to 'lengths', each row contains
a length observation and columns represent discrete observation units, such as year.
When 'dataType' is set to 'freq', each row is the number of length observations corresponding to a length bin.
The first column must contain the mid-points of each length bin, with columns 2 and above representing discrete observation units
}
