% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishLengthAssess.R
\docType{data}
\name{LengthCompExample}
\alias{LengthCompExample}
\title{Length composition example}
\format{
An S4 object of class LengthComp
}
\source{
Simulated data
}
\usage{
LengthCompExample
}
\description{
A LengthComp object, prepopulated with an example simulated from LBSPRsim.
}
\details{
The life history used to generate lenth composition is based on the Beverton-Holt life history invariants of M/K=1.5 and Lm/Linf=0.66
}
\keyword{datasets}
