% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishLengthAssess.R
\docType{data}
\name{LengthCompExampleLength}
\alias{LengthCompExampleLength}
\title{Length composition example - raw length}
\format{
An S4 object of class LengthComp
}
\source{
Simulated data
}
\usage{
LengthCompExampleLength
}
\description{
A LengthComp object using Lengths, populated with an example simulated from LBSPRsim.
}
\details{
The life history used to generate lenth composition is based on the Beverton-Holt life history invariants of M/K=1.5 and Lm/Linf=0.66
}
\keyword{datasets}
