% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{LoptFunc}
\alias{LoptFunc}
\title{Optimum harvest length Beverton (1992)}
\usage{
LoptFunc(LifeHistoryObj)
}
\arguments{
\item{LifeHistoryObj}{A life history object.}
}
\description{
Beverton, R.J.H., 1992. Patterns of reproductive strategy parameters in some marine teleost fishes. J. Fish. Biol. 41, 137-160.
}
\examples{
library(fishSimGTG)
LoptFunc(fishSimGTG::LifeHistoryExample)
}
