% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_size_bins}
\alias{make_size_bins}
\title{Build SizeBins list from a LifeHistory object
Creates the SizeBins list required by GTG dome-shaped LBSPR functions.
Uses CVLinf (0.1) and MaxSD (2) if not set as attributes on the LifeHistory object
CVLinf(0.1) and maxSD(2) ensure the length vector always extends beyond Linf to capture the full length distribution}
\usage{
make_size_bins(lifeHistoryObj, Linc = 1)
}
\arguments{
\item{lifeHistoryObj}{A LifeHistory S4 object with Linf defined}

\item{Linc}{Numeric. Length bin width in cm (default = 1).}
}
\description{
Build SizeBins list from a LifeHistory object
Creates the SizeBins list required by GTG dome-shaped LBSPR functions.
Uses CVLinf (0.1) and MaxSD (2) if not set as attributes on the LifeHistory object
CVLinf(0.1) and maxSD(2) ensure the length vector always extends beyond Linf to capture the full length distribution
}
\examples{
library(fishSimGTG)
lh <- fishSimGTG::LifeHistoryExample
SizeBins <- make_size_bins(lh)
}
