% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsAssess.R
\name{PLcFunc}
\alias{PLcFunc}
\title{Proportion of catch above the minumum size limit}
\usage{
PLcFunc(Lc, LengthCompObj, byGroup = FALSE)
}
\arguments{
\item{Lc}{A minimum size limit in the same units of measure (L_units) and type of measurement (L_type) as the corresponding LengthComp object.}

\item{LengthCompObj}{A LengthComp object}

\item{byGroup}{A logical indicating whether quantity is to be calculated separately for each of multiple length comp groups (TRUE) or to length comp is to be pooled across groups prior to calculating quantity (default = FALSE). When TRUE, pooling is ignored if only a single group exists.}
}
\description{
Proportion of the length observations that are equal to or greater than a minimum size limit
}
\examples{
library(fishSimGTG)
PLcFunc(50, fishLengthAssess::LengthCompExample, byGroup = FALSE)
}
