% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_gillnet_doc.R
\docType{data}
\name{raw_data_gillnet}
\alias{raw_data_gillnet}
\title{Example gillnet catch data}
\format{
A data frame with 11 rows and 9 variables:
\describe{
\item{V1}{Mid-point length}
\item{V2}{Catch for mesh size 1}
\item{V3}{Catch for mesh size 2}
\item{V4}{Catch for mesh size 3}
\item{V5}{Catch for mesh size 4}
\item{V6}{Catch for mesh size 5}
\item{V7}{Catch for mesh size 6}
\item{V8}{Catch for mesh size 7}
\item{V9}{Catch for mesh size 8}
}
}
\source{
Adapted from SELECT and TropFishR
}
\usage{
raw_data_gillnet
}
\description{
This dataset contains example gillnet catch data across multiple mesh sizes (eight mesh sizes).
It is adapted from SELECT and TropFishR and used to demonstrate dome-shaped selectivity fitting.
First column: MidLength.
Second - ninth columns: Mesh sizes.
}
\keyword{datasets}
