% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishLengthAssess.R
\docType{package}
\name{fishLengthAssess}
\alias{fishLengthAssess-package}
\alias{fishLengthAssess}
\title{Length-based evaluation of fish stock and fishery}
\description{
Evaluates stock status and fishery compliance, and conducts other management-related tasks based on length-frequency data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/natureanalytics-ca/fishLengthAssess}
  \item Report bugs at \url{https://github.com/natureanalytics-ca/fishLengthAssess/issues}
}

}
\author{
\strong{Maintainer}: William Harford \email{bill@natureanalytics.ca}

}
